--- spec/helpers/admin/videos_helper_spec.rb
@@ -1,15 +1,0 @@
-require 'spec_helper'
-
-# Specs in this file have access to a helper object that includes
-# the Admin::VideosHelper. For example:
-#
-# describe Admin::VideosHelper do
-#   describe "string concat" do
-#     it "concats two strings with spaces" do
-#       expect(helper.concat_strings("this","that")).to eq("this that")
-#     end
-#   end
-# end
-describe Admin::VideosHelper do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
